feat: make useDB custom hook
useDB 라는 커스텀 훅을 만들어, Body 컴포넌트와 커플링을 느슨하게 만듦.
